# backend/cmd/server/openapi.yaml
openapi: 3.0.3
info:
  title: TODO API
  description: A simple API for managing TODO items.
  version: 1.0.0

paths:
  /api/todos:
    get:
      summary: List all TODO items
      description: Retrieves a list of TODO items, optionally filtered by date.
      parameters:
        - name: date
          in: query
          description: The date to filter TODOs by (YYYY-MM-DD).
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: A list of TODOs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
    post:
      summary: Create a new TODO item
      description: Adds a new item to the TODO list.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTodo'
      responses:
        '201':
          description: The created TODO item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'

  /api/todos/{id}:
    patch:
      summary: Update a TODO item
      description: Updates an existing TODO item's text or completion status.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodo'
      responses:
        '200':
          description: The updated TODO item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
    delete:
      summary: Delete a TODO item
      description: Deletes a TODO item by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '24':
          description: No Content. The item was deleted successfully.

components:
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: string
          example: "todo-12345"
        text:
          type: string
          example: "Buy milk"
        completed:
          type: boolean
          example: false
        date:
          type: string
          format: date
          example: "2025-11-25"
    NewTodo:
      type: object
      required:
        - text
        - date
      properties:
        text:
          type: string
          example: "Learn OpenAPI"
        date:
          type: string
          format: date
          example: "2025-11-26"
    UpdateTodo:
      type: object
      properties:
        text:
          type: string
          example: "Learn Go and OpenAPI"
        completed:
          type: boolean
          example: true